backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.8
pydantic-settings==2.2.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
sqlalchemy==2.0.23
alembic==1.12.1
# databases[sqlite]==0.8.0  # 注释掉冲突的包
redis==5.0.1
httpx==0.25.2
aiofiles==23.2.1
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.2
yfinance==0.2.33
plotly==5.18.0
matplotlib==3.8.2
scipy==1.11.4
scikit-learn==1.2.2
chromadb==0.4.22
sentence-transformers==2.2.2
openai==1.3.9
tiktoken==0.5.2
loguru==0.7.2
tencentcloud-sdk-python==3.0.1120
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.9.1
pytz==2023.3.post1
tzdata==2023.3
python-dateutil==2.8.2
Jinja2==3.1.2
MarkupSafe==2.1.3
weasyprint==61.0
reportlab==4.0.7
PyPDF2==3.0.1
qrcode[pil]==7.4.2
pillow==10.1.0